Move exception handler to utils
